% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.HOmics.R
\name{filter.HOmics}
\alias{filter.HOmics}
\title{filters results using parameter thresholds}
\usage{
\method{filter}{HOmics}(res, param = "p.pos", threshold = 0.95,
  as.data.frame = T)
}
\arguments{
\item{res}{object of class HOmics, obtained from function HOmics}

\item{param}{probability to filter with p.pos and p.neg}

\item{threshold}{numerical, the threshold to select associated variables related to the specified param. Default is 0.95}

\item{as.data.frame}{collapse as data.frame. Default = TRUE}
}
\description{
filters results using parameter thresholds
}
