% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.signif.R
\name{get.signif}
\alias{get.signif}
\title{filters results using p-value thresholds}
\usage{
get.signif(res, param = "p.pos", threshold = 0.05, as.data.frame = T)
}
\arguments{
\item{res}{resultslist}

\item{param}{probability to filter with p.pos and p.neg}

\item{threshold}{numerical, the threshold to select associated variables related to the specified param. Default is 0.05}

\item{as.data.frame}{collapse as data.frame. Default = TRUE}
}
\description{
filters results using p-value thresholds
}
