% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsignif.R
\name{getsignif}
\alias{getsignif}
\title{filters results using p-value thresholds}
\usage{
getsignif(res, param = "p.pos", threshold = 0.05, as.data.frame = T)
}
\arguments{
\item{res}{resultslist}

\item{param}{parameter of the table to filter with (vars_n, aic, Cor2, p or adj.p). Default is "adj.p"}

\item{threshold}{numerical, the threshold to select associated variables related to the specified param. Default is 0.05}

\item{as.data.frame}{collapse as data.frame. Default = TRUE}

\item{comp}{comparison param comp threshold. Default is "below"}
}
\description{
filters results using p-value thresholds
}
