% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOmics.meth.R
\name{HOmics.meth}
\alias{HOmics.meth}
\title{Integrates methylation and gene data}
\usage{
HOmics.meth(meth.data = GRSet, pheno.cond.col = "tissue.ch1",
  annot.gene.col = "UCSC_RefGene_Name",
  annot.z.col = c("UCSC_RefGene_Group"), annot.mult.sep = ";",
  z.matrix = NULL, pheno.covar.col = NULL, gene.list = genes.u,
  seed = NULL, cores = 1, ...)
}
\arguments{
\item{meth.data}{GenomicRatioSet or  ExpressionSet with methylation data}

\item{pheno.cond.col}{response variable, column of annotation that contains condition (one of the variable names contained in pheno data obtained with function pData())}

\item{annot.gene.col}{column of annotation that contains gene names (same annotations as gene.list). Default = "UCSC_RefGene_Name"}

\item{annot.z.col}{prior information. Column of annotation that contains prior variable (for GenomicRatioSet functions getAnnotation() or fData() for ExpressionSet). Default = "UCSC_RefGene_Group"}

\item{annot.mult.sep}{separator for annot.gene.col and maybe annot.z.col multiple values. Default = ";"}

\item{z.matrix}{prior information. Alternative parameter to annot.z.col as matrix with rows the featureNames of meth.data and columns the Zvars can be given. Default = NULL}

\item{pheno.covar.col}{covariates, column of annotation that contains names of covariates to include in the model (one or more of the variable names contained in pheno data). Default = NULL}

\item{gene.list}{genes to analyze as a vector with symbols (HUGO)}

\item{seed}{numerical seed for the use of function set.seed in the generation of the model, for reproducibility}

\item{cores}{cores in case of parallelization. Default = 1 (no parallelization)}
}
\value{
an object of class HOmics
}
\description{
Integrates methylation and gene data
}
\examples{
to be built
}
