% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOmics.R
\name{HOmics}
\alias{HOmics}
\title{Integrates transcritpomics and methylation}
\usage{
HOmics(data.matrix, agg.matrix, cond, z.matrix, cores = 1, ...)
}
\arguments{
\item{data.matrix}{matrix with rownames the features and columns the samples}

\item{agg.matrix}{matrix with colnames the features and rows the groups according to some feature aggregation criteria, 0 for non pertenance}

\item{cond}{response variable, usually a numerical factor with two levels representing the conditions to compare. If cond is a numerical vector (continuous response), a hiearchical linear regression model will be fit instead of the default hierarchical logistic regression model}

\item{z.matrix}{column of annotation that contains gene names (same annotations as gene.list)}

\item{cores}{cores in case of parallelization. Default=1}
}
\value{
an object of class HOmics
}
\description{
Integrates transcritpomics and methylation
}
\examples{
to be built
}
