% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{integrate}
\alias{integrate}
\title{Integrates transcritpomics and methylation}
\usage{
integrate(meth.data = GRSet, pheno.cond.col = "tissue.ch1",
  annot.gene.col = "UCSC_RefGene_Name",
  annot.z.cols = c("UCSC_RefGene_Group"), annot.mult.sep = NULL,
  zvars = NULL, gene.list = genes.u, cores = 1, ...)
}
\arguments{
\item{meth.data}{GenomicRatioSet o multidataset o expressionset}

\item{pheno.cond.col}{response variable, one of the variable names contained in pData of meth.data}

\item{annot.z.cols}{prior information. vector of variable names contained in featureData (for GenomicRatioSet functions getAnnotation(brge_methy) or granges(brge_methy) could be used).}

\item{zvars}{prior information. Alternative parameter to annot.z.cols as matrix with rows the featureNames of meth.data and columns the Zvars can be given. Default=NULL}

\item{cores}{cores in case of parallelization. Default=1}
}
\value{
integrate
}
\description{
Integrates transcritpomics and methylation
}
\examples{
to be built
}
