% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.HOmics.R
\name{plot.HOmics}
\alias{plot.HOmics}
\title{plots results in terms of 95\% credible interval}
\usage{
\method{plot}{HOmics}(res, element = 1)
}
\arguments{
\item{res}{object of class HOmics, obtained from function HOmics}

\item{element}{number or name of group of features to plot. Default = 1}
}
\description{
plots results in terms of 95\% credible interval
}
