% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.res.R
\name{plot.res}
\alias{plot.res}
\title{plots results in terms of 95CI}
\usage{
\method{plot}{res}(res, element = 1)
}
\arguments{
\item{res}{resultslist}

\item{element}{number or name of group of features to plot. Default = 1}
}
\description{
plots results in terms of 95CI
}
