% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.HOmics.R
\name{print.HOmics}
\alias{print.HOmics}
\title{Print for an object of class HOmics}
\usage{
\method{print}{HOmics}(res)
}
\arguments{
\item{res}{an HOmics object as obtained in HOmics or HOmics.meth()}
}
\description{
Print for an object of class HOmics
}
