% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif.HOmics.R
\name{signif.HOmics}
\alias{signif.HOmics}
\title{returns only significative results (containing no 0s)}
\usage{
\method{signif}{HOmics}(res, as.data.frame = T)
}
\arguments{
\item{res}{object of class HOmics, obtained from function HOmics}

\item{as.data.frame}{collapse as data.frame. Default = TRUE}
}
\description{
returns only significative results (containing no 0s)
}
