% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmodel.R
\name{hmodel}
\alias{hmodel}
\title{Performs rjags models}
\usage{
hmodel(g.matrix, z.matrix, cond.v)
}
\arguments{
\item{g.matrix}{matrix of cpgs, with samples in rows and cpgs in columns}

\item{z.matrix}{matrix of functional annotations for g.matrix cpgs, with cpgs in rows and functions in columns}

\item{cond.v}{response variable, as vector}
}
\value{
hmodel
}
\description{
Performs rjags models
}
\examples{
to be built
}
