% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmodel.R
\name{hmodel}
\alias{hmodel}
\title{Performs rjags models}
\usage{
hmodel(g.matrix, z.matrix, cond, cont, covar.matrix = NULL, seed)
}
\arguments{
\item{g.matrix}{matrix of cpgs, with samples in rows and cpgs in columns}

\item{z.matrix}{matrix of functional annotations for g.matrix cpgs, with cpgs in rows and functions in columns}

\item{cond}{response variable, as numerical vector}

\item{cont}{flag to consider a continuous response}

\item{covar.matrix}{matrix of continuous covariates (in columns)}
}
\value{
results for a gene model
}
\description{
Performs rjags models
}
\examples{
to be built
}
