% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOmics.R
\name{HOmics}
\alias{HOmics}
\title{Integrates two omic data through hierarchical modeling}
\usage{
HOmics(data.matrix, agg.matrix, cond, z.matrix, cores = 1, ...)
}
\arguments{
\item{data.matrix}{matrix with rownames the features and columns the samples}

\item{agg.matrix}{matrix with colnames the features and rows the groups according to some feature aggregation criteria, 0 for non pertenance}

\item{cond}{response variable, usually a numerical factor with two levels representing the conditions to compare. If cond is a numerical vector (continuous response), a hiearchical linear regression model will be fit instead of the default hierarchical logistic regression model}

\item{z.matrix}{matrix with prior information related to features, with rownames the features and columns the samples}

\item{cores}{cores in case of parallelization. Default = 1 (no parallelization)}
}
\value{
an object of class HOmics
}
\description{
Integrates two omic data through hierarchical modeling
}
\examples{
to be built
}
