% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOmics.R
\name{HOmics}
\alias{HOmics}
\title{Integrates two omic data through hierarchical modeling}
\usage{
HOmics(data.matrix, agg.matrix, cond, z.matrix, covar.matrix = NULL,
  seed = NULL, cores = 1, ...)
}
\arguments{
\item{data.matrix}{matrix with features as rownames and samples as columns}

\item{agg.matrix}{matrix with features as rownames and columns corresponding to the groups, according to some feature aggregation criteria, 0 for non pertenance}

\item{cond}{response variable, usually a numerical factor with two levels representing the conditions to compare. If cond is a numerical vector (continuous response), a hiearchical linear regression model will be fit instead of the default hierarchical logistic regression model}

\item{z.matrix}{matrix with prior information related to features, with rownames the features and columns the samples}

\item{covar.matrix}{vector or matrix of continuous covariates, with samples as rownames (in the same order as cond) and covariates as columns. Default = NULL}

\item{seed}{numerical seed for the use of function set.seed in the generation of the model, for reproducibility}

\item{cores}{cores in case of parallelization. Default = 1 (no parallelization)}
}
\value{
an object of class HOmics
}
\description{
Integrates two omic data through hierarchical modeling
}
\examples{
to be built
}
